refactor: major backend cleanup — single DB pool, int64 ASN, split router, fix module path
- Consolidate all DB pool creation into cmd/api/main.go; stores accept *pgxpool.Pool
- NewServer returns (*Server, error) — no partial servers on failure
- Replace goroutine-per-prefix farm with single unnest() CTE query per table
- Split router.go into per-feature files (prefix, asn, sets, datasources, caching)
- Extract writeJSON into internal/httputil, delete 3 duplicates
- Fix cache-hit JSON double-encode (write raw bytes on HIT)
- Use net.SplitHostPort in rate limiter for correct IPv6 handling
- ASN as int64 across all domain types, stores, irrd client, datasources
- Delete dead redirect handlers (asPath, whois, advancedSearch)
- Delete navigation package (no schema, no UI wiring, always 500'd)
- Replace prefixSummaryAlias with type-alias MarshalJSON pattern
- Remove prefixSortKey* fields from wire format
- Repo hygiene: delete dead project files, move PNG to docs/
- Frontend: replace bare catch{} with proper error categorization + AbortController
- Replace hand-rolled containsInt/containsString with slices.Contains
- Expose per-RIR freshness in /api/metadata/, surface RIR import failures
- Fix module path: github.com/sebastiaan/irrexplorer -> gitlab.int.koetsier.org/sebas/irrexplorer

+++ go-backend/internal/datasources/peeringdb.go
@@ -21,7 +21,7 @@ func NewPeeringDBClient(timeout time.Duration) *PeeringDBClient {
 	}
 }
 
-func (c *PeeringDBClient) QueryASN(ctx context.Context, asn int) map[string]any {
+func (c *PeeringDBClient) QueryASN(ctx context.Context, asn int64) map[string]any {
 	data, err := c.get(ctx, fmt.Sprintf("%s/net?asn=%d", c.baseURL, asn))
 	if err != nil {
 		return map[string]any{"asn": asn, "error": err.Error()}
